+++ packages/wesley-cli/src/commands/init.mjs
@@ -59,4 +59,3 @@ export class InitCommand extends WesleyCommand {
   }
 }
 
-