+++ packages/wesley-cli/src/commands/rehearse.mjs
@@ -145,4 +145,3 @@ async function execSql(db, dsn, sql) {
   return db.query(dsn, sql);
 }
 
-