+++ packages/wesley-cli/src/commands/transform.mjs
@@ -21,4 +21,3 @@ export class TransformPipelineCommand extends WesleyCommand {
   }
 }
 
-